feat: byteWriter light weight, uses little memory good for goroutines
